Finally fix chaining bug
Can now chain any amount of operations. Either by
operating on the answer, or by pressing a number
and starting over - no need to clear! It will also
show the previous answer until you submit the new
operation! Additionally I added an error message for trying
to divide by 0 and started working on a 'correct' button.
the CSS and HTML have been added but as I am having issues
it will take a but longer.