Get correct button working
Works before and after an answer is added. And correctly
clears the view and the stored value. Returns to null
when the string hits '' and doesn't break any further
calculations. Going to try and figure out this box shadow
issue and may come back in the future to resolve it. But
everything works for now on the JS side.