Get shadows working
Realized I didn't understand the syntax which was
the issue here. Also shadows on the button did not
look good so i removed that. But on the calc it
makes it look 3D so i kept it. Also added fonts and
a nice background for the page so its not such a large
contrast. Am going to move on from this now i cant see
much else to add.